chore(backend): rename dtos to dto
