+++ apps/backend/src/modules/auth/dtos/index.ts
@@ -1,6 +1,6 @@
-export * from './common.dtos';
-export * from './verifications/request-verification.dtos';
-export * from './verifications/verify-verification.dtos';
-export * from './verifications/inspect-verification.dtos';
-export * from './password/change-password.dtos';
-export * from './password/reset-password.dtos';
+export * from './common.dto';
+export * from './verifications/request-verification.dto';
+export * from './verifications/verify-verification.dto';
+export * from './verifications/inspect-verification.dto';
+export * from './password/change-password.dto';
+export * from './password/reset-password.dto';
