+++ apps/backend/src/modules/auth/dtos/password/change-password.dto.ts
