+++ apps/backend/src/modules/auth/dtos/verifications/inspect-verification.dto.ts
@@ -1,6 +1,6 @@
 import { t } from 'elysia';
 import { createDtoSchemas } from '~backend/shared/middlewares/dto/schemas';
-import { verificationDtoSchema } from '../common.dtos';
+import { verificationDtoSchema } from '../common.dto';
 
 export const inspectVerificationDtoSchemas = createDtoSchemas({
 	params: t.Object({
