+++ apps/backend/src/modules/users/dtos/get-user-textures.dto.ts
@@ -1,6 +1,6 @@
 import { t } from 'elysia';
 import { createDtoSchemas } from '~backend/shared/middlewares/dto/schemas';
-import { userTextureResponseSchema } from './common.dtos';
+import { userTextureResponseSchema } from './common.dto';
 
 export const getUserTexturesDtoSchemas = createDtoSchemas({
 	params: t.Object({
