+++ apps/backend/src/modules/users/dtos/index.ts
@@ -1,4 +1,4 @@
-export * from './common.dtos';
-export * from './get-user-info.dtos';
-export * from './get-user-profiles.dtos';
-export * from './get-user-textures.dtos';
+export * from './common.dto';
+export * from './get-user-info.dto';
+export * from './get-user-profiles.dto';
+export * from './get-user-textures.dto';
